+++ tests/rider_test.rs
@@ -32,6 +32,7 @@ fn cmd(head: &str, args: &[&str]) -> Cmd {
         by_reference: false,
         env_assigns: Default::default(),
         is_intra_command_function: false,
+        expandable_args: Default::default(),
     }
 }